@@ -9,7 +9,7 @@ public class OneOrMore : Pattern, IHasChildren
 
     public override IEnumerable<INode> ChildNodes => this.Children;
 
-    public IPattern[] Children => this.Self.Elements().Select(this.ToPattern).ToArray();
+    public IPattern[] Children => [.. this.Self.Elements().Select(this.ToPattern)];
 
     internal static OneOrMore Parse(RngElement element, SchemaContext context)
     {
