@@ -7,7 +7,7 @@ public class Ref : Pattern
     {
     }
 
-    public override IEnumerable<INode> ChildNodes => Array.Empty<Node>();
+    public override IEnumerable<INode> ChildNodes => [];
 
     public string Name => this.Self.Attribute("name").Value;
 
